#98 [Rename] 패키지명 변경
controller -> api
프로젝트 내에서 restAPI만 취급하기 때문에 변경
