+++ src/main/java/com/project/storemanager_api/api/AuthController.java
@@ -1,4 +1,4 @@
-package com.project.storemanager_api.controller;
+package com.project.storemanager_api.api;
 
 import com.project.storemanager_api.domain.user.dto.request.LoginRequestDto;
 import com.project.storemanager_api.domain.user.dto.request.ModifyUserRequestDto;
