+++ src/main/java/com/project/storemanager_api/api/MenuController.java
@@ -1,4 +1,4 @@
-package com.project.storemanager_api.controller;
+package com.project.storemanager_api.api;
 
 
 import com.project.storemanager_api.domain.menu.dto.request.ModifyMenuRequestDto;
