+++ src/main/java/com/project/storemanager_api/api/PlaceController.java
@@ -1,4 +1,4 @@
-package com.project.storemanager_api.controller;
+package com.project.storemanager_api.api;
 
 import com.project.storemanager_api.domain.place.dto.request.ModifyPlaceRequestDto;
 import com.project.storemanager_api.domain.place.dto.request.SavePlaceRequestDto;
